Move locationsFor to lib
